Minor bug fix: Call correct time conversion
Swapped calls to getTimeStrFromDays and getTimeStrFromMs so that when selected from the context menu, Time (given in days) results in the number being interpreted in days and Time (given in ms) results in the number being interpreted in milliseconds.  Before, these were reversed so Time (given in days) resulted in the number being interpreted in milliseconds and Time (given in ms) resulted in the number being interpreted in days.